requirements.txt
setuptools == 3.3
GitPython == 1.0.2
Requests == 2.9.1
